Refactor different functionality into different modules.
